# client/openapi.json
{"openapi":"3.1.0","info":{"title":"Image Clustering API","description":"An API for uploading, processing, and clustering images.","version":"1.0.0"},"paths":{"/images/upload":{"post":{"tags":["Images"],"summary":"Upload Images","description":"Uploads one or more image files and triggers background processing for each.","operationId":"upload_images_images_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_images_images_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ImageResponse"},"type":"array","title":"Response Upload Images Images Upload Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/images/":{"get":{"tags":["Images"],"summary":"Get All Images","description":"Retrieves a list of all images in the database.","operationId":"get_all_images_images__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ImageResponse"},"type":"array","title":"Response Get All Images Images  Get"}}}}}}},"/images/{image_id}":{"get":{"tags":["Images"],"summary":"Get Image File","description":"Returns the original image file.","operationId":"get_image_file_images__image_id__get","parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"integer","title":"Image Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Images"],"summary":"Delete Image","description":"Deletes an image's database record and its physical files.","operationId":"delete_image_images__image_id__delete","parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"integer","title":"Image Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/images/thumbnail/{image_id}":{"get":{"tags":["Images"],"summary":"Get Thumbnail File","description":"Returns the thumbnail file for an image.","operationId":"get_thumbnail_file_images_thumbnail__image_id__get","parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"integer","title":"Image Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/batches/":{"get":{"tags":["Clustering Batches"],"summary":"Get All Batches","description":"Retrieves a list of all clustering batches.","operationId":"get_all_batches_batches__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/BatchResponse"},"type":"array","title":"Response Get All Batches Batches  Get"}}}}}},"post":{"tags":["Clustering Batches"],"summary":"Create Batch","description":"Creates a new, unprocessed batch record with a list of image IDs.","operationId":"create_batch_batches__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/batches/{batch_id}":{"get":{"tags":["Clustering Batches"],"summary":"Get Batch Details","description":"Gets all information about a specific batch by its ID.","operationId":"get_batch_details_batches__batch_id__get","parameters":[{"name":"batch_id","in":"path","required":true,"schema":{"type":"integer","title":"Batch Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/batches/{batch_id}/images":{"post":{"tags":["Clustering Batches"],"summary":"Add Images To Batch","description":"Adds one or more images to an existing batch.","operationId":"add_images_to_batch_batches__batch_id__images_post","parameters":[{"name":"batch_id","in":"path","required":true,"schema":{"type":"integer","title":"Batch Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchUpdateImages"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Clustering Batches"],"summary":"Remove Images From Batch","description":"Removes one or more images from an existing batch.","operationId":"remove_images_from_batch_batches__batch_id__images_delete","parameters":[{"name":"batch_id","in":"path","required":true,"schema":{"type":"integer","title":"Batch Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchUpdateImages"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/batches/{batch_id}/analyze":{"put":{"tags":["Clustering Batches"],"summary":"Analyze Batch","description":"Runs (or re-runs) the clustering analysis on a batch.","operationId":"analyze_batch_batches__batch_id__analyze_put","parameters":[{"name":"batch_id","in":"path","required":true,"schema":{"type":"integer","title":"Batch Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchAnalyze"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Read Root","description":"A simple health check endpoint.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BatchAnalyze":{"properties":{"eps":{"type":"number","title":"Eps","default":0.3},"min_samples":{"type":"integer","title":"Min Samples","default":2},"metric":{"type":"string","title":"Metric","default":"cosine"}},"type":"object","title":"BatchAnalyze","description":"Schema for the analysis request."},"BatchCreate":{"properties":{"name":{"type":"string","title":"Name"},"image_ids":{"items":{"type":"integer"},"type":"array","title":"Image Ids"}},"type":"object","required":["name","image_ids"],"title":"BatchCreate","description":"Schema for the initial creation of a batch."},"BatchResponse":{"properties":{"id":{"type":"integer","title":"Id"},"batch_name":{"type":"string","title":"Batch Name"},"status":{"type":"string","title":"Status"},"image_ids":{"items":{"type":"integer"},"type":"array","title":"Image Ids"},"parameters":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Parameters"},"cluster_summary":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Cluster Summary"}},"type":"object","required":["id","batch_name","status","image_ids","parameters","cluster_summary"],"title":"BatchResponse","description":"Schema for returning full batch details."},"BatchUpdateImages":{"properties":{"image_ids":{"items":{"type":"integer"},"type":"array","title":"Image Ids"}},"type":"object","required":["image_ids"],"title":"BatchUpdateImages","description":"Schema for adding/removing images from a batch."},"Body_upload_images_images_upload_post":{"properties":{"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files"}},"type":"object","required":["files"],"title":"Body_upload_images_images_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ImageResponse":{"properties":{"id":{"type":"integer","title":"Id"},"filename":{"type":"string","title":"Filename"},"original_filename":{"type":"string","title":"Original Filename"},"file_path":{"type":"string","title":"File Path"},"has_thumbnail":{"type":"boolean","title":"Has Thumbnail"}},"type":"object","required":["id","filename","original_filename","file_path","has_thumbnail"],"title":"ImageResponse","description":"Schema for returning image details."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}